https://github.com/ikawrakow/ik_llama.cpp/pull/71 have a good idea.

I'll figure to add it in llamafile/tinyblas sgemm (and a litle more) and id work great:

- AMD Ryzen 9 7940HS (zen4)

> Mistral-Nemo-Instruct-2407.BF16.gguf +kv@bf16

| Model          | Test   |   t/s master |   t/s perfo/tinyblas |   Speedup |
|:---------------|:-------|-------------:|---------------------:|----------:|
| llama 13B BF16 | pp1    |         2.52 |                 2.51 |      1.00 |
| llama 13B BF16 | pp2    |         5.00 |                 4.74 |      0.95 |
| llama 13B BF16 | pp3    |         7.44 |                 7.23 |      0.97 |
| llama 13B BF16 | pp4    |         9.75 |                 9.91 |      1.02 |
| llama 13B BF16 | pp5    |        11.95 |                12.37 |      1.04 |
| llama 13B BF16 | pp6    |        13.92 |                14.78 |      1.06 |
| llama 13B BF16 | pp7    |        15.64 |                17.09 |      1.09 |
| llama 13B BF16 | pp8    |        17.24 |                19.41 |      1.13 |
| llama 13B BF16 | pp9    |        18.35 |                21.63 |      1.18 |
| llama 13B BF16 | pp10   |        19.47 |                24.02 |      1.23 |
| llama 13B BF16 | pp11   |        20.48 |                26.30 |      1.28 |
| llama 13B BF16 | pp12   |        21.04 |                28.43 |      1.35 |
| llama 13B BF16 | pp13   |        21.49 |                29.41 |      1.37 |
| llama 13B BF16 | pp14   |        23.10 |                31.56 |      1.37 |
| llama 13B BF16 | pp15   |        23.65 |                33.51 |      1.42 |
| llama 13B BF16 | pp16   |        23.99 |                35.87 |      1.50 |
| llama 13B BF16 | pp30   |        24.19 |                51.09 |      2.11 |
| llama 13B BF16 | pp32   |        24.56 |                51.04 |      2.08 |
| llama 13B BF16 | pp64   |        25.66 |                57.23 |      2.23 |
| llama 13B BF16 | pp65   |        24.57 |                57.33 |      2.33 |
| llama 13B BF16 | pp120  |        25.73 |                65.51 |      2.55 |
| llama 13B BF16 | pp128  |        25.77 |                65.81 |      2.55 |
| llama 13B BF16 | pp130  |        26.86 |                66.31 |      2.47 |
| llama 13B BF16 | pp240  |        27.76 |                70.40 |      2.54 |
| llama 13B BF16 | pp255  |        26.03 |                70.32 |      2.70 |
| llama 13B BF16 | pp256  |        26.04 |                70.34 |      2.70 |
| llama 13B BF16 | pp510  |        25.73 |                68.26 |      2.65 |
| llama 13B BF16 | pp512  |        25.74 |                67.97 |      2.64 |
| llama 13B BF16 | pp1024 |        25.27 |                66.76 |      2.64 |
| llama 13B BF16 | pp1025 |        25.04 |                64.84 |      2.59 |
| llama 13B BF16 | pp2048 |        24.63 |                63.96 |      2.60 |
| llama 13B BF16 | tg128  |         2.52 |                 2.52 |      1.00 |

> Mistral-Nemo-Instruct-2407.FP16.gguf +kv@fp16

| Model         | Test   |   t/s master |   t/s perfo/tinyblas |   Speedup |
|:--------------|:-------|-------------:|---------------------:|----------:|
| llama 13B F16 | pp1    |         2.50 |                 2.50 |      1.00 |
| llama 13B F16 | pp2    |         4.94 |                 4.81 |      0.97 |
| llama 13B F16 | pp3    |         7.41 |                 7.19 |      0.97 |
| llama 13B F16 | pp4    |         9.81 |                 9.82 |      1.00 |
| llama 13B F16 | pp5    |        12.23 |                12.25 |      1.00 |
| llama 13B F16 | pp6    |         7.92 |                14.58 |      1.84 |
| llama 13B F16 | pp7    |         9.19 |                16.52 |      1.80 |
| llama 13B F16 | pp8    |        10.47 |                18.76 |      1.79 |
| llama 13B F16 | pp9    |        11.76 |                20.86 |      1.77 |
| llama 13B F16 | pp10   |        22.58 |                22.80 |      1.01 |
| llama 13B F16 | pp11   |        14.19 |                24.90 |      1.75 |
| llama 13B F16 | pp12   |        15.41 |                26.70 |      1.73 |
| llama 13B F16 | pp13   |        16.66 |                26.99 |      1.62 |
| llama 13B F16 | pp14   |        17.88 |                28.74 |      1.61 |
| llama 13B F16 | pp15   |        31.97 |                29.67 |      0.93 |
| llama 13B F16 | pp16   |        19.79 |                31.23 |      1.58 |
| llama 13B F16 | pp30   |        38.31 |                36.86 |      0.96 |
| llama 13B F16 | pp32   |        29.11 |                36.60 |      1.26 |
| llama 13B F16 | pp64   |        32.15 |                38.95 |      1.21 |
| llama 13B F16 | pp65   |        38.72 |                38.91 |      1.00 |
| llama 13B F16 | pp120  |        39.14 |                40.36 |      1.03 |
| llama 13B F16 | pp128  |        35.44 |                40.19 |      1.13 |
| llama 13B F16 | pp130  |        39.49 |                40.24 |      1.02 |
| llama 13B F16 | pp240  |        36.90 |                40.76 |      1.10 |
| llama 13B F16 | pp255  |        35.87 |                40.66 |      1.13 |
| llama 13B F16 | pp256  |        33.51 |                40.43 |      1.21 |
| llama 13B F16 | pp510  |        27.96 |                40.09 |      1.43 |
| llama 13B F16 | pp512  |        27.41 |                40.08 |      1.46 |
| llama 13B F16 | pp1024 |        27.27 |                39.03 |      1.43 |
| llama 13B F16 | pp1025 |        25.91 |                38.50 |      1.49 |
| llama 13B F16 | pp2048 |        26.75 |                37.95 |      1.42 |
| llama 13B F16 | tg128  |         2.50 |                 2.51 |      1.00 |



- on AMD Ryzen 9 5950X 16-Core Processor (znver3) (AVX2)
> Mistral-Nemo-Instruct-2407.BF16.gguf +kv@bf16

| Model          | Test   |   t/s master |   t/s perfo/tinyblas |   Speedup |
|:---------------|:-------|-------------:|---------------------:|----------:|
| llama 13B BF16 | pp1    |         2.21 |                 2.21 |      1.00 |
| llama 13B BF16 | pp2    |         4.36 |                 4.31 |      0.99 |
| llama 13B BF16 | pp3    |         6.44 |                 6.44 |      1.00 |
| llama 13B BF16 | pp4    |         8.42 |                 8.58 |      1.02 |
| llama 13B BF16 | pp5    |        10.29 |                10.71 |      1.04 |
| llama 13B BF16 | pp6    |        12.00 |                12.78 |      1.07 |
| llama 13B BF16 | pp7    |        13.53 |                14.86 |      1.10 |
| llama 13B BF16 | pp8    |        14.72 |                16.92 |      1.15 |
| llama 13B BF16 | pp9    |        15.61 |                18.93 |      1.21 |
| llama 13B BF16 | pp10   |        16.30 |                20.92 |      1.28 |
| llama 13B BF16 | pp11   |        16.93 |                23.01 |      1.36 |
| llama 13B BF16 | pp12   |        17.35 |                24.89 |      1.43 |
| llama 13B BF16 | pp13   |        17.69 |                26.94 |      1.52 |
| llama 13B BF16 | pp14   |        17.95 |                28.78 |      1.60 |
| llama 13B BF16 | pp15   |        18.21 |                30.64 |      1.68 |
| llama 13B BF16 | pp16   |        18.37 |                32.45 |      1.77 |
| llama 13B BF16 | pp30   |        19.20 |                42.87 |      2.23 |
| llama 13B BF16 | pp32   |        19.36 |                43.14 |      2.23 |
| llama 13B BF16 | pp64   |        19.85 |                45.05 |      2.27 |
| llama 13B BF16 | pp65   |        19.46 |                44.94 |      2.31 |
| llama 13B BF16 | pp120  |        19.98 |                46.27 |      2.32 |
| llama 13B BF16 | pp128  |        20.14 |                46.11 |      2.29 |
| llama 13B BF16 | pp130  |        19.97 |                45.93 |      2.30 |
| llama 13B BF16 | pp240  |        20.23 |                46.50 |      2.30 |
| llama 13B BF16 | pp255  |        20.24 |                46.54 |      2.30 |
| llama 13B BF16 | pp256  |        20.19 |                46.40 |      2.30 |
| llama 13B BF16 | pp510  |        20.09 |                46.01 |      2.29 |
| llama 13B BF16 | pp512  |        20.17 |                45.81 |      2.27 |
| llama 13B BF16 | pp1024 |        19.94 |                45.05 |      2.26 |
| llama 13B BF16 | pp1025 |        19.74 |                44.18 |      2.24 |
| llama 13B BF16 | pp2048 |        19.48 |                43.68 |      2.24 |
| llama 13B BF16 | tg128  |         2.21 |                 2.21 |      1.00 |

> Mistral-Nemo-Instruct-2407.FP16.gguf +kv@fp16

| Model         | Test   |   t/s master |   t/s perfo/tinyblas |   Speedup |
|:--------------|:-------|-------------:|---------------------:|----------:|
| llama 13B F16 | pp1    |         2.19 |                 2.19 |      1.00 |
| llama 13B F16 | pp2    |         4.30 |                 4.28 |      1.00 |
| llama 13B F16 | pp3    |         6.46 |                 6.41 |      0.99 |
| llama 13B F16 | pp4    |         4.84 |                 8.53 |      1.76 |
| llama 13B F16 | pp5    |         6.01 |                10.64 |      1.77 |
| llama 13B F16 | pp6    |        12.90 |                12.71 |      0.99 |
| llama 13B F16 | pp7    |         8.50 |                14.81 |      1.74 |
| llama 13B F16 | pp8    |         9.64 |                16.88 |      1.75 |
| llama 13B F16 | pp9    |        19.25 |                18.90 |      0.98 |
| llama 13B F16 | pp10   |        12.25 |                20.88 |      1.70 |
| llama 13B F16 | pp11   |        13.39 |                22.94 |      1.71 |
| llama 13B F16 | pp12   |        25.40 |                24.89 |      0.98 |
| llama 13B F16 | pp13   |        15.89 |                26.87 |      1.69 |
| llama 13B F16 | pp14   |        17.02 |                28.74 |      1.69 |
| llama 13B F16 | pp15   |        30.82 |                30.66 |      0.99 |
| llama 13B F16 | pp16   |        19.45 |                32.55 |      1.67 |
| llama 13B F16 | pp30   |        34.23 |                54.23 |      1.58 |
| llama 13B F16 | pp32   |        27.34 |                55.37 |      2.03 |
| llama 13B F16 | pp64   |        30.66 |                58.46 |      1.91 |
| llama 13B F16 | pp65   |        31.03 |                57.47 |      1.85 |
| llama 13B F16 | pp120  |        35.31 |                58.01 |      1.64 |
| llama 13B F16 | pp128  |        33.09 |                57.76 |      1.75 |
| llama 13B F16 | pp130  |        33.05 |                58.28 |      1.76 |
| llama 13B F16 | pp240  |        35.19 |                58.66 |      1.67 |
| llama 13B F16 | pp255  |        35.24 |                58.62 |      1.66 |
| llama 13B F16 | pp256  |        33.98 |                58.57 |      1.72 |
| llama 13B F16 | pp510  |        33.76 |                57.94 |      1.72 |
| llama 13B F16 | pp512  |        33.34 |                57.51 |      1.72 |
| llama 13B F16 | pp1024 |        33.03 |                56.41 |      1.71 |
| llama 13B F16 | pp1025 |        32.59 |                53.96 |      1.66 |
| llama 13B F16 | pp2048 |        32.08 |                54.93 |      1.71 |
| llama 13B F16 | tg128  |         2.18 |                 2.19 |      1.00 |