@slaren 2024-12-11
On my system (intel 13900k) I see better performance with BF16, but worse with F16 in some cases:

| Model         | Test   |   t/s master |   t/s perfo/tinyblas |   Speedup |
|:--------------|:-------|-------------:|---------------------:|----------:|
| llama 7B BF16 | pp32   |        33.94 |                42.27 |      1.25 |
| llama 7B BF16 | pp64   |        34.82 |                43.27 |      1.24 |
| llama 7B BF16 | pp128  |        33.22 |                43.69 |      1.32 |
| llama 7B BF16 | tg32   |         6.75 |                 6.41 |      0.95 |
| llama 7B F16  | pp32   |        41.45 |                28.85 |      0.70 |
| llama 7B F16  | pp64   |        42.88 |                26.34 |      0.61 |
| llama 7B F16  | pp128  |        43.69 |                29.75 |      0.68 |
| llama 7B F16  | tg32   |         6.82 |                 6.47 |      0.95 |

With different numbers of threads:
| Model        |   Threads | Test   |   t/s master |   t/s perfo/tinyblas |   Speedup |
|:-------------|----------:|:-------|-------------:|---------------------:|----------:|
| llama 7B F16 |         8 | pp64   |        51.98 |                59.14 |      1.14 |
| llama 7B F16 |        16 | pp64   |        35.20 |                28.97 |      0.82 |
| llama 7B F16 |        24 | pp64   |        63.18 |                43.40 |      0.69 |
| llama 7B F16 |        32 | pp64   |        75.45 |                54.18 |      0.72 |